Ground query to rely on raycast and spherecast instead of sweep
It might be useful to be able to configure this the cast origin to allow
for a "step" height.

Sweep would be ideal, but Rigidbody.Sweeptest fails if the object is too
close to the object that you're expecting to pick up in the sweep.

This forces our character controller to utilize a CapsuleCollider for
now, but I think this is an acceptable constraint if it means we can
make good headway on this project.